requirements.txt
absl-py==0.1.12
astor==0.6.2
backports.weakref==1.0rc1
bleach==1.5.0
enum34==1.1.6
gast==0.2.0
grpcio==1.10.0
html5lib==0.9999999
Markdown==2.2.0
numpy==1.14.2
Pillow==5.0.0
protobuf==3.5.2.post1
scipy==1.0.1
six==1.11.0
tensorboard==1.6.0
tensorflow-gpu==1.2.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
Werkzeug==0.14.1
